Added in the Firestore Synchronization. Calibration of Data will follow the following Structure.
Each individual Access Point will have their own Document under the Collection "Access Points".
Each Access Point Document will be labeled according to their BSSID and contain maps to their BSSID and SSID for ease of access purposes. The Document will also contain a map of all Locations that have detected this Access Point and will store the relative RSSI value in this map with the Location as the Key.

Thus, when calibrating, for each access point that appears on the screen, the app will access the document referencing that access point and update the map with the Current Location and RSSI detected at Current Location.

Note: The calibrating proccess is inefficient with this data structure but the query process is a magnitude more efficient because of this.
Signed-off-by: Mudaafi <Mudaafi@gmail.com>