Implemented two functions: queryFirestore and localizationAlgorithm
In queryFirestore, I used the code that runs synchronously with pulling data from Firestore. This was done using a while-loop to test for task completion which may put an intense load on the client device. However, asynchronous pulling seems a little dangerous as it may override some data. Also, I haven't figured out how to pause until all the asynchronous pulling is completed which is really tiring and head-scratching. In any case, queryFirestore will now pull data stored in Firestore and convert it from the store DS of AP -> Location : RSSI to Location --> AP : RSSI as requested.

In localizationAlgorithm, I mostly copypasted the code you sent me, adjusting it to fit a hash-in-a-hash DS rather than an array because I needed to use a map when converting from Firestore. Upon calling, it will now display the minimum value, the location and the final bssid on the screen.

On a final note, I think the code rn is all over the place and not properly abstracted so I'll try to work on that in the future. Also,  used a lot of (what I think are) global variables but honestly I don't know how to do without them because of the asynchronous calls which can't seem to return anything. An option is to use a callback function but I haven't properly researched on those yet.

Signed-off-by: Mudaafi <Mudaafi@gmail.com>