Implemented 1 function: getUserLocation Implemented AlertDialog which is basically a pop-up input field to allow us to change the location variable to whatever want. Needed change for calibration later on. I call "getUserLocation" in the calibration function "getAndScanResult" as well. This AlertDialog is Asynchronous which is a bit of a problem but I finally managed to implement a work around using a public interface "UserStringCallback". Intense stuff.
Implemented  1 User Interface: UserStringCallback
A callback interface that is used in getUserLocation to synchronize the user-input dialog with the running function.

Abstracted getAndScanResults with pushCalibration.
"pushCalibration" is essentially all the code that pushes the data from ClientEnd to Firestore.

As a final note, I still need to do a lot of abstraction and documentation plus implement a sync for the async query from firestore. God that's  gonna be such a pain. Will keep u updated.

Signed-off-by: Mudaafi <Mudaafi@gmail.com>